% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPhrase.R
\name{checkPhrase}
\alias{checkPhrase}
\title{Check some text for a phrase}
\usage{
checkPhrase(phrase, issueBody)
}
\arguments{
\item{phrase}{The phrase to look for.}

\item{issueBody}{The text to search within.}
}
\value{
A logical vector.
}
\description{
Check some text for a phrase
}
\examples{
checkPhrase('hello', 'you say hello i say goodbye')
}
