% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuetriager-package.R
\docType{package}
\name{issuetriager-package}
\alias{issuetriager}
\alias{issuetriager-package}
\title{issuetriager: Automated Triage of Support Issues}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Elliot Murphy \email{elliot@rstudio.com} (\href{https://orcid.org/0000-0001-7473-8081}{ORCID})

Other contributors:
\itemize{
  \item  RStudio, PBC [copyright holder, funder]
}

}
\keyword{internal}
